# services/order-service/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Order Service API",
    "description": "Order Service API documentation",
    "version": "1.0"
  },
  "host": "localhost:8083",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/orders": {
      "post": {
        "summary": "Create order",
        "description": "Create a new order",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "orders"
        ],
        "parameters": [
          {
            "description": "Order Data",
            "name": "order",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateOrderRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Order created successfully",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "get": {
        "summary": "List user orders",
        "description": "List orders for a specific user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "orders"
        ],
        "responses": {
          "200": {
            "description": "List of user orders",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          }
        }
      }
    },
    "/orders/:id/status": {
      "put": {
        "summary": "Update order status",
        "description": "Update the status of an existing order",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "orders"
        ],
        "parameters": [
          {
            "type": "integer",
            "description": "Order ID",
            "name": "order_id",
            "in": "path",
            "required": true
          },
          {
            "description": "New Status",
            "name": "status",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateStatusRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Order status updated successfully",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/order-items": {
      "post": {
        "summary": "Create order item",
        "description": "Add a new item to an order",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "order-items"
        ],
        "parameters": [
          {
            "description": "Order Item Data",
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateOrderItemRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Order item created successfully",
            "schema": {
              "$ref": "#/definitions/OrderItem"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/order-items/:id": {
      "get": {
        "summary": "Get order item",
        "description": "Get a specific order item by ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "order-items"
        ],
        "parameters": [
          {
            "type": "integer",
            "description": "Order Item ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Order item details",
            "schema": {
              "$ref": "#/definitions/OrderItem"
            }
          },
          "404": {
            "description": "Order item not found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "put": {
        "summary": "Update order item",
        "description": "Update quantity and total price of an order item",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "order-items"
        ],
        "parameters": [
          {
            "type": "integer",
            "description": "Order Item ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Updated Item Data",
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateOrderItemRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order item updated successfully",
            "schema": {
              "$ref": "#/definitions/OrderItem"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Order item not found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "delete": {
        "summary": "Delete order item",
        "description": "Delete an item from an order",
        "produces": [
          "application/json"
        ],
        "tags": [
          "order-items"
        ],
        "parameters": [
          {
            "type": "integer",
            "description": "Order Item ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Order item deleted successfully",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Order item not found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/orders/:order_id/items": {
      "get": {
        "summary": "List order items",
        "description": "Get all items for a specific order",
        "produces": [
          "application/json"
        ],
        "tags": [
          "order-items"
        ],
        "parameters": [
          {
            "type": "integer",
            "description": "Order ID",
            "name": "order_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "List of order items",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OrderItem"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "user_id": {
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "note": {
          "type": "string"
        },
        "order_items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          }
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        },
        "deleted_at": {
          "type": "string"
        }
      }
    },
    "OrderItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "order_id": {
          "type": "integer"
        },
        "product_id": {
          "type": "integer"
        },
        "product_name": {
          "type": "string"
        },
        "product_slug": {
          "type": "string"
        },
        "variant_id": {
          "type": "integer"
        },
        "variant_sku": {
          "type": "string"
        },
        "variant_price": {
          "type": "number"
        },
        "quantity": {
          "type": "integer"
        },
        "total_price": {
          "type": "number"
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        },
        "deleted_at": {
          "type": "string"
        }
      }
    },
    "CreateOrderRequest": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "integer"
        },
        "note": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          }
        }
      }
    },
    "CreateOrderItemRequest": {
      "type": "object",
      "required": [
        "order_id",
        "product_id",
        "product_name",
        "product_slug",
        "variant_id",
        "variant_sku",
        "variant_price",
        "quantity",
        "total_price"
      ],
      "properties": {
        "order_id": {
          "type": "integer"
        },
        "product_id": {
          "type": "integer"
        },
        "product_name": {
          "type": "string"
        },
        "product_slug": {
          "type": "string"
        },
        "variant_id": {
          "type": "integer"
        },
        "variant_sku": {
          "type": "string"
        },
        "variant_price": {
          "type": "number"
        },
        "quantity": {
          "type": "integer"
        },
        "total_price": {
          "type": "number"
        }
      }
    },
    "UpdateOrderItemRequest": {
      "type": "object",
      "required": [
        "quantity",
        "total_price"
      ],
      "properties": {
        "quantity": {
          "type": "integer"
        },
        "total_price": {
          "type": "number"
        }
      }
    },
    "UpdateStatusRequest": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "in_progress"
        }
      }
    }
  }
}